chore: move EditDevicePopop to gui/device module
