+++ src/network_puzzles/gui/device.kv
@@ -9,3 +9,87 @@
     DeviceLabel:
         text: root.hostname
         pos_hint: {'center_x': 0.5}
+
+<EditDevicePopup>:
+    size_hint: (0.75, 0.6)
+    ThemedBoxLayout:
+        canvas.before:
+            Color:
+                rgba: (1, 1, 1, 1)
+            Rectangle:
+                pos: self.pos
+                size: self.size
+        spacing: dp(20)
+        ThemedBoxLayout:
+            size_hint_x: 0.3
+            orientation: "vertical"
+            ThemedBoxLayout:
+                size_hint_y: 0.16
+                size_hint_max_y: root.app.BUTTON_MAX_H
+                ThemedLabel:  # Hostname
+                    text: _("Hostname")
+                TextInput:  # device.hostname
+                    multiline: False
+                    text: root.device.hostname
+            PopupButton:  # Gateway
+                size_hint_y: 0.28
+                text: _("Gateway")
+                on_press: root.on_gateway()
+            PopupButton:  # Routes
+                size_hint_y: 0.28
+                text: _("Routes")
+                on_press: root.on_routes()
+            PopupButton:  # VLANs
+                size_hint_y: 0.28
+                text: _("VLANs")
+                on_press: root.on_vlans()
+            Widget:
+                size_hint_y: 0.28
+        ThemedBoxLayout:
+            size_hint_x: 0.3
+            orientation: "vertical"
+            ThemedBoxLayout:
+                size_hint_y: 0.15
+                ThemedLabel:  # NICs
+                    size_hint_x: 0.45
+                    text: _("NICs")
+                PopupButton:  # Edit
+                    size_hint_x: 0.55
+                    text: _("Edit")
+                    on_press: root.on_nics_edit()
+            TextInput:  # available NICs
+                size_hint_y: 0.50
+                text: root.get_editable_nic_names()
+            ThemedBoxLayout:
+                size_hint_y: 0.15
+                ThemedLabel:  # IPs
+                    size_hint_x: 0.15
+                    text: _("IPs")
+                PopupButton:  # +
+                    size_hint_x: 0.15
+                    text: "+"
+                    on_press: root.on_ips_add()
+                PopupButton:  # -
+                    size_hint_x: 0.15
+                    text: "-"
+                    on_press: root.on_ips_remove()
+                PopupButton:  # Edit
+                    size_hint_x: 0.55
+                    text: _("Edit")
+                    on_press: root.on_ips_edit()
+            ThemedLabel:
+                size_hint_y: 0.20
+                text: "###.###.###.###"
+        ThemedBoxLayout:
+            size_hint_x: 0.4
+            orientation: "vertical"
+            TextInput:  # ARP table
+                text: "XXXXXXXX"
+            ThemedBoxLayout:
+                orientation: "vertical"
+                PopupButton:  # Cancel
+                    text: _("Cancel")
+                    on_press: root.dismiss()
+                PopupButton:  # OK
+                    text: _("Okay")
+                    on_press: root.on_okay()
